requirements.txt
requests
beautifulsoup4
playwright
faiss-cpu
sentence-transformers
streamlit
